+++ CHANGELOG.rst
@@ -35,8 +35,10 @@ Unreleased
 * Added package __version__.
 * Added routine for getting any network component from key.    
 * Added bus function get_v_max(code) get_v_min(code).
-  
-  
+* Added output_level to network show_components.
+* Added unittest for redundant buses.
+* Fixed build_lib.sh to handle PWD with white spaces.
+
 Version 1.3.3
 -------------
 * Made Python wrapper its own repo.